+++ docs/changelog.md
@@ -2,6 +2,10 @@
 
 This changelog documents user-relevant changes to the tmate SSH server charm.
 
+## 2025-10-14
+
+- Fixed the charm status not reflecting errors with starting the tmate SSH daemon.
+
 ## 2025-08-20
 
 - Added changelog to track user relevant changes.